refactor(lattices): Rename `bottom.rs` -> `with_bot.rs`, `top.rs` -> `with_top.rs` 1/4 (#763)
