+++ lattices/src/with_bot.rs
